requirements.txt
# Core Web Framework
flask==3.1.0
python-dotenv==1.1.0
werkzeug==3.1.3

# Data Processing
pandas==2.2.3
openpyxl==3.1.5

# AI Models - Large Language Models (Text + Vision Domain)
google-generativeai==0.8.4
mistralai>=1.0.0,<2.0.0
httpx>=0.27.0  # HTTP client for Ollama local LLM API

# PDF Processing
pdfplumber==0.11.6
pypdf2==3.0.1
PyMuPDF>=1.23.0

# OCR - Tesseract (Image Domain - CNN-based text detection)
pytesseract>=0.3.10
Pillow>=10.0.0

# Document Classification (Image Classification Domain - Vision Transformer)
transformers>=4.30.0
torch>=2.0.0

# Speech-to-Text - Whisper (Audio Domain - Transformer encoder-decoder)
openai-whisper>=20231117

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
